Second try: Attempting to resolve build issue with bogus repository described by jira.codehaus.org/browse/MEV-649
Rather than reverting to an older version of log4j, I'm declaring
central as the first repository in the root pom.  This will hopefully
have the effect of convincing maven to check this repository before
any other repository when doing a build.

In addition, I made sure all existing repositories in the root
pom had snapshots disabled.  We don't currently have any SNAPSHOT
dependencies (and shouldn't, given how close we are to the 3.3 release).


git-svn-id: svn://svn.code.sf.net/p/fedora-commons/code/fedora/trunk@8369 6cc838c6-e627-0410-8cce-a6db3432f928